feat: 컨트롤러/파사드 검색 구현
