+++ src/main/java/com/example/musinssak/api/product/facade/ProductFacade.java
@@ -1,0 +1,4 @@
+package com.example.musinssak.api.product.facade;
+
+public class ProductFacade {
+}
