+++ src/main/java/com/example/musinssak/api/product/facade/ProductFacadeImpl.java
@@ -1,0 +1,4 @@
+package com.example.musinssak.api.product.facade;
+
+public class ProductFacadeImpl {
+}
