+++ src/main/java/com/example/musinssak/common/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.example.musinssak.config;
+package com.example.musinssak.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
